Added and moved Structs properly
