+++ go-backend/structs/DeviceWithAppsDTO.go
@@ -22,13 +22,3 @@ type ApplicationInstanceDTO struct {
 	Status      string `json:"status"`
 	Path        string `json:"path"`
 }
-
-// Define the Tag domain model
-// TODO:move to seperate file
-type Tag struct {
-	ID         int64  `json:"id"`
-	Name       string `json:"name"`
-	Type       string `json:"type"`
-	IsEditable bool   `json:"is_editable"`
-	OwnerID    *int64 `json:"owner_id"` // Nullable foreign key
-}
